* Changed typedefs of uint32, uint64, etc to use the correct ones under both microsoft and gnu compilers. (__int32, and on *nix int32_t) - Partially fixed DBC code under big endian (this shit needs a rewrite)

+++ src/shared/Network/SocketOpsWin32.cpp
@@ -21,14 +21,14 @@ namespace SocketOps
 	// Disable blocking send/recv calls.
 	bool Nonblocking(SOCKET fd)
 	{
-		uint32 arg = 1;
+		u_long arg = 1;
 		return (::ioctlsocket(fd, FIONBIO, &arg) == 0);
 	}
 
 	// Disable blocking send/recv calls.
 	bool Blocking(SOCKET fd)
 	{
-		uint32 arg = 0;
+		u_long arg = 0;
 		return (ioctlsocket(fd, FIONBIO, &arg) == 0);
 	}
